Fix code snippets
