+++ examples/docs_snippets/docs_snippets/integrations/dbt/quickstart/with_project.py
@@ -1,48 +1,6 @@
 # pyright: reportMissingImports=false
 # ruff: isort: skip_file
 
-# start_example
-from pathlib import Path
-
-from dagster import AssetExecutionContext, Definitions
-from dagster_dbt import (
-    DbtCliResource,
-    DbtProject,
-    build_schedule_from_dbt_selection,
-    dbt_assets,
-)
-
-RELATIVE_PATH_TO_MY_DBT_PROJECT = "./my_dbt_project"
-
-my_project = DbtProject(
-    project_dir=Path(__file__)
-    .joinpath("..", RELATIVE_PATH_TO_MY_DBT_PROJECT)
-    .resolve(),
-)
-
-
-@dbt_assets(manifest=my_project.manifest_path)
-def my_dbt_assets(context: AssetExecutionContext, dbt: DbtCliResource):
-    yield from dbt.cli(["build"], context=context).stream()
-
-
-my_schedule = build_schedule_from_dbt_selection(
-    [my_dbt_assets],
-    job_name="materialize_dbt_models",
-    cron_schedule="0 0 * * *",
-    dbt_select="fqn:*",
-)
-
-defs = Definitions(
-    assets=[my_dbt_assets],
-    schedules=[my_schedule],
-    resources={
-        "dbt": DbtCliResource(project_dir=my_project),
-    },
-)
-# end_example
-
-
 # start_dbt_project_example
 from pathlib import Path
 
@@ -81,7 +39,7 @@ from .project import my_project
 
 defs = Definitions(
     ...,
-    assets=[
+    assets=[  # type: ignore
         ...,
         # Add the dbt assets alongside your other asset
         my_dbt_assets,
